# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Stardust API
  version: "0.1.0"
  description: >
    HTTP/JSON API over the Stardust core. Every route maps to a core operation
    also reachable from the CLI and the MCP server. Bind to localhost; this API
    has no authentication and is intended for local/NAS use behind your own trust
    boundary.
servers:
  - url: http://127.0.0.1:7777
paths:
  /healthz:
    get:
      summary: Liveness check
      responses:
        "200":
          description: OK
  /query:
    get:
      summary: Hybrid keyword + semantic search
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QueryResult" }
        "400": { description: Missing q }
  /note:
    get:
      summary: Read a parsed note by vault-relative path
      parameters:
        - name: path
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Parsed note
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Note" }
        "404": { description: Not found }
  /status:
    get:
      summary: Index health and configuration
      responses:
        "200":
          description: Status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Status" }
  /graph:
    get:
      summary: Derive the link graph and report orphans + broken links
      responses:
        "200":
          description: Graph report
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GraphReport" }
  /cron:
    get:
      summary: List declarative cron jobs
      responses:
        "200": { description: Jobs }
  /mounts:
    get:
      summary: List configured external-source mounts
      responses:
        "200":
          description: Mounts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Mount" }
  /collections:
    get:
      summary: List configured collections with live record counts
      responses:
        "200":
          description: Collections
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CollectionInfo" }
  /collection:
    get:
      summary: Read a single collection by name
      parameters:
        - name: name
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Collection
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CollectionInfo" }
        "400": { description: Missing name }
        "404": { description: Not found }
  /records:
    get:
      summary: List records in a collection, filtered and sorted
      parameters:
        - name: collection
          in: query
          required: true
          schema: { type: string }
        - name: where
          in: query
          description: >
            Frontmatter filter as field:op:value, repeatable. op is one of eq,
            ne, gt, gte, lt, lte, contains. Filters combine with AND.
          schema: { type: array, items: { type: string } }
        - name: sort
          in: query
          description: >
            Sort field (a frontmatter key, or path / updated_at), prefix with -
            for descending.
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer }
        - name: offset
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Record list
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RecordList" }
        "400": { description: Missing collection or malformed where }
    post:
      summary: Create a record in a collection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [collection]
              properties:
                collection: { type: string }
                fields: { type: object, additionalProperties: true }
                body: { type: string }
      responses:
        "200":
          description: Created record
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Record" }
        "400": { description: Missing collection or malformed body }
  /record:
    get:
      summary: Read a single record by vault-relative path
      parameters:
        - name: path
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Record
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Record" }
        "400": { description: Missing path }
        "404": { description: Not found }
    patch:
      summary: Patch a record's frontmatter and/or body
      parameters:
        - name: path
          in: query
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fields:
                  type: object
                  additionalProperties: true
                  description: >
                    Frontmatter columns to merge; a null value deletes that key.
                body: { type: string }
      responses:
        "200":
          description: Updated record
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Record" }
        "400": { description: Missing path or malformed body }
  /index:
    post:
      summary: Incrementally index the vault
      parameters:
        - name: since
          in: query
          schema: { type: string, description: git SHA for the diff fast path }
      responses:
        "200":
          description: Index stats
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IndexStats" }
  /rebuild:
    post:
      summary: Nuke the cache and reindex from scratch
      responses:
        "200":
          description: Index stats
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IndexStats" }
  /archive:
    post:
      summary: Snapshot git history into a destination folder
      parameters:
        - name: dest
          in: query
          schema: { type: string }
      responses:
        "200": { description: "{ path }" }
  /cron/run:
    post:
      summary: Run a cron job now
      parameters:
        - name: job
          in: query
          required: true
          schema: { type: string }
      responses:
        "200": { description: "{ output }" }
        "400": { description: Missing job }
components:
  schemas:
    Hit:
      type: object
      properties:
        path: { type: string }
        title: { type: string }
        heading: { type: string }
        snippet: { type: string }
        score: { type: number }
    QueryResult:
      type: object
      properties:
        query: { type: string }
        mode: { type: string }
        hits:
          type: array
          items: { $ref: "#/components/schemas/Hit" }
    Note:
      type: object
      properties:
        path: { type: string }
        title: { type: string }
        tags: { type: array, items: { type: string } }
        links: { type: array, items: { type: string } }
        link_targets:
          type: array
          items:
            type: object
            properties:
              link: { type: string }
              path: { type: string }
        body: { type: string }
    Mount:
      type: object
      properties:
        name: { type: string }
        kind: { type: string }
        target: { type: string }
        args: { type: array, items: { type: string } }
        tool: { type: string }
    Status:
      type: object
      properties:
        root: { type: string }
        notes: { type: integer }
        chunks: { type: integer }
        last_indexed_sha: { type: string }
        embed_model: { type: string }
        vectors: { type: boolean }
        reranker: { type: boolean }
    GraphReport:
      type: object
      properties:
        notes: { type: integer }
        links: { type: integer }
        orphans: { type: array, items: { type: string } }
        broken:
          type: array
          items:
            type: object
            properties:
              from: { type: string }
              target: { type: string }
        pagerank:
          type: array
          items:
            type: object
            properties:
              path: { type: string }
              title: { type: string }
              score: { type: number }
    IndexStats:
      type: object
      properties:
        indexed: { type: integer }
        skipped: { type: integer }
        deleted: { type: integer }
        vectors: { type: boolean }
    Field:
      type: object
      properties:
        name: { type: string }
        type:
          type: string
          enum: [string, number, bool, date, enum, tags, ref]
        required: { type: boolean }
        enum: { type: array, items: { type: string } }
        default: {}
    CollectionInfo:
      type: object
      properties:
        name: { type: string }
        path: { type: string, description: vault-relative folder the records live in }
        description: { type: string }
        fields:
          type: array
          items: { $ref: "#/components/schemas/Field" }
        records: { type: integer, description: live record count }
    Record:
      type: object
      properties:
        path: { type: string }
        title: { type: string }
        frontmatter: { type: object, additionalProperties: true }
        body: { type: string }
    RecordList:
      type: object
      properties:
        collection: { type: string }
        folder: { type: string }
        records:
          type: array
          items: { $ref: "#/components/schemas/Record" }